Move AuthContext into context dir
