+++ client/package.json
@@ -20,6 +20,7 @@
     "react-markdown": "^6.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "react-use-websocket": "^2.7.1",
     "remark-gfm": "^1.0.0",
     "source-map-explorer": "^2.5.2",
     "swr": "^0.5.6",
