+++ client/src/LoggedOutView/Forms.js
@@ -2,7 +2,7 @@ import React from "react"
 import { Link, useHistory, useLocation } from "react-router-dom"
 import { Formik, Form, Field } from "formik"
 
-import AuthContext from "../AuthContext.js"
+import AuthContext from "../context/AuthContext.js"
 import { BASE_URL } from "../utils/api.js"
 
 import { Input, Button } from "./FormComponents.js"
